refactor: reuse `IncompleteTopologyException`
