+++ gateway/src/main/java/io/camunda/zeebe/gateway/admin/IncompleteTopologyException.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.gateway.admin.backup;
+package io.camunda.zeebe.gateway.admin;
 
 public class IncompleteTopologyException extends RuntimeException {
